# docs/swagger.yaml
basePath: /
definitions:
...
  code.Failure:
    properties:
      code:
        description: 业务码
        type: integer
      data:
        description: 业务返回数据
      message:
        description: 描述信息
        type: string
    type: object
...
  tool.hashIdsDecodeResponse:
    properties:
      val:
        description: 解密后的值
        type: integer
    type: object
...
  tool.searchCacheResponse:
    properties:
      ttl:
        description: 过期时间
        type: string
      val:
        description: 查询后的值
        type: string
    type: object
  tool.searchMySQLResponse:
    properties:
      cols:
        description: 查询后的行
        items:
          type: string
        type: array
      cols_info:
        description: 行信息
        items:
          $ref: '#/definitions/tool.tableColumn'
        type: array
      list:
        description: 查询后的数据
        items:
          additionalProperties: true
          type: object
        type: array
    type: object
...
  tool.tableColumn:
    properties:
      column_comment:
        description: 字段注释
        type: string
      column_name:
        description: 字段名称
        type: string
    type: object
...
info:
  contact: {}
  license:
    name: MIT
    url: https://github.com/ChangSZ/gin-boilerplate/blob/master/LICENSE
  title: swagger 接口文档
  version: "2.0"
paths:
...
  /api/tool/cache/search:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 查询缓存
      parameters:
      - description: Redis Key
        in: formData
        name: redis_key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tool.searchCacheResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - LoginToken: []
      summary: 查询缓存
      tags:
      - API.tool
...
  /api/tool/data/mysql:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 执行 SQL 语句
      parameters:
      - description: 数据库名称
        in: formData
        name: db_name
        required: true
        type: string
      - description: 数据表名称
        in: formData
        name: table_name
        required: true
        type: string
      - description: SQL 语句
        in: formData
        name: sql
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tool.searchMySQLResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - LoginToken: []
      summary: 执行 SQL 语句
      tags:
      - API.tool
...
  /api/tool/hashids/decode/{id}:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: HashIds 解密
      parameters:
      - description: 需解密的密文
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tool.hashIdsDecodeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/code.Failure'
      security:
      - LoginToken: []
      summary: HashIds 解密
      tags:
      - API.tool
...
securityDefinitions:
  LoginToken:
    in: header
    name: token
    type: apiKey
swagger: "2.0"